% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanSet}
\alias{cleanSet}
\title{Get the Question Set and Objective Set Scores}
\usage{
cleanSet(HWsheet, key)
}
\arguments{
\item{HWsheet}{The output of function \code{cleanScore} $HWsheet}

\item{key}{The output of function \code{convertKey}}
}
\value{
List of two data frames of scores for the question
sets and learning outcomes respectively.
}
\description{
Get the Question Set and Objective Set Scores
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

