% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataAnalysis.R
\name{clusterCriteria}
\alias{clusterCriteria}
\title{Calculate the criterion values for different number of clusters}
\usage{
clusterCriteria(dat, ntopic, l = 9)
}
\arguments{
\item{dat}{Output of function \code{cl}}

\item{ntopic}{Number of topics}

\item{l}{Number of clusters}
}
\value{
Data frame of three columns: cluster, criteria, and values
}
\description{
The criteria include SSR, SSE, and R-squared
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

