% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAnalysis.R
\name{clusterStudents}
\alias{clusterStudents}
\title{Cluster students by performance}
\usage{
clusterStudents(dat, l = 9)
}
\arguments{
\item{dat}{Data frame with rows being the students and
columns being the scores over the semester}

\item{l}{Maximum number of clusters to try}
}
\value{
a Data frame with the left part being the scaled
\code{dat} and the right part being the clustering labels
for 1:\code{l} clusters
}
\description{
Cluster students by performance
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

