% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCSV.R
\name{combineFiles}
\alias{combineFiles}
\title{Merge two CSV files into one file}
\usage{
combineFiles(file1, file2, newnm)
}
\arguments{
\item{file1}{File name with path to the first file}

\item{file2}{File name with path to the second file}

\item{newnm}{File name for the returned merged file}
}
\value{
A new CSV file
}
\description{
If one topic is organized into two chapters, then we must
merge together the two files. Note that the Question ID and RsQ ID
are updated for the second file upon merging.
}
\examples{
library(ePort)

##### *. Split data files #####
dataPath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Data Files/Ch3"
namelist = list.files(path=dataPath,full.names=TRUE)
for (i in namelist) rewriteData(i) # May not need this rewriting step!
for(i in c('AB','CD','EF','GH','JKQ')){
  tmp = namelist[grep(i,basename(namelist))]
  splitFile(tmp, 29, "ID")
}

##### *. Merge data files #####
dataPath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Fall 2014/Data Files/Ch9-Ch11"
namelist = list.files(path=dataPath,full.names=TRUE)
for (i in namelist) rewriteData(i) # May not need this rewriting step!
for(i in c('AB','CD','EF','GH','JK','LM')){
  tmp = namelist[grep(i,basename(namelist))]
  combineFiles(tmp[2],tmp[1],paste("Topic11",i,"csv",sep='.'))
}
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

