% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_key.R
\name{convertkey}
\alias{convertkey}
\title{Convert a TXT answer key to a data frame}
\usage{
convertkey(filenm)
}
\arguments{
\item{filenm}{File name of the answer key with path}
}
\value{
List of two data frames. The rows of the first data frame
are the answers for each question. The rows of the second data frame are the questions.
}
\description{
Answer keys downloaded from Respondus are plain txt files. This function will convert them into data frames.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}
\seealso{
\url{https://www.respondus.com/}
}

