% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/de-identify_student.R
\name{encodeName}
\alias{encodeName}
\title{Encode all student names in the data files}
\usage{
encodeName(dpath, dict = "nameCode.csv")
}
\arguments{
\item{dpath}{Path to the data files}

\item{dict}{File name with path for the dictionary. Default is "nameCode.csv"}
}
\description{
Read the name dictionary, and rewrite the data files
into the de-identified names
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

