% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merge_sections.R
\name{merge_section}
\alias{merge_section}
\title{Summarise several files of the same topic}
\usage{
merge_section(filenmlist, answerkey, skip = NULL)
}
\arguments{
\item{filenmlist}{a vector of file names of the Blackboard data with path}

\item{answerkey}{the output of function \code{convertkey}}

\item{skip}{a vector of integers. Same as the parameter in \code{clean_score}.}
}
\value{
a list of results including the summaries and merged data
}
\description{
There are usually 6-7 sections each semester. This function
is used to extract the scores from different sections and merge
them together.
}
\examples{
library(ePort)
library(plyr)
library(dplyr)

##### save data files into one csv file #####
key_txt = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Database Questions/Topic05.Questions.txt"
datapath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Data Files"
topic = "05"
namelist = list.files(path=datapath,full.names=TRUE)
namelist = namelist[grep(paste('Topic',topic,'\\\\.',sep=''),basename(namelist))]
namelist

answerkey = convertkey(key_txt)
instructor_scores = merge_section(namelist, answerkey,skip=NULL)
rownames(instructor_scores$score) = instructor_scores$score$student
tmp1 = instructor_scores$Qscore
l = ncol(tmp1)
tmp1$TotalScore = rowSums(tmp1, na.rm=TRUE)
tmp1$Section = instructor_scores$score[rownames(tmp1),'Section']
tmp1 = tmp1[,c(l+2,l+1,1:l)]
tmp2 = instructor_scores$Qanswer
tmp2$Section = instructor_scores$score[rownames(tmp2),'Section']
tmp2 = tmp2[,c(l+1,1:l)]
write.csv(tmp1,file='Topic3.score.csv')
write.csv(tmp2,file='Topic3.answer.csv')
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}
