% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSections.R
\name{mergeSection}
\alias{mergeSection}
\title{Summarise several files of the same topic}
\usage{
mergeSection(filenmlist, answerkey, skip = NULL)
}
\arguments{
\item{filenmlist}{a vector of file names of the Blackboard data with path}

\item{answerkey}{the output of function \code{convertKey}}

\item{skip}{a vector of integers. Same as the parameter in \code{cleanScore}.}
}
\value{
a list of results including the summaries and merged data
}
\description{
There are usually 6-7 sections each semester. This function
is used to extract the scores from different sections and merge
them together.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

