% Please edit documentation in R/clean_key.R
\name{refine_key}
\alias{refine_key}
\title{Parse the html file to get the answer key.}
\usage{
refine_key(key)
}
\arguments{
\item{key}{path of the .htm file}
}
\value{
the cleaned answer key in plain text format.
}
\description{
Parse the html file to get the answer key.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}
