% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanKey.R
\name{refineKey}
\alias{refineKey}
\title{Parse the html file to get the answer key.}
\usage{
refineKey(key)
}
\arguments{
\item{key}{path of the .htm file}
}
\value{
the cleaned answer key in plain text format.
}
\description{
Parse the html file to get the answer key.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

