% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_csv.R
\name{rewrite_data}
\alias{rewrite_data}
\title{Rewrite the data file}
\usage{
rewrite_data(filenm)
}
\arguments{
\item{filenm}{file name with path}
}
\description{
Solve two problems that I used to work in gedit:
"" --> ", and "\\r\\n" --> \\r\\n.
This function will read the data, replace every two quotes ""
by one quote ", remove the quotes before and after every line,
i.e., replace "\\r\\n" by \\r\\n, then write the result to the
same file.
}
\examples{
library(ePort)
library(knitr)

##### Homework reports #####
## key ##
# save the .doc file as .htm format, to keep the images
key_htm = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Database Questions/Topic03.Questions.htm"
refine_key(key_htm) # generate a clean answer key with paths to the plots in the question
key_txt = gsub("htm$","txt",key_htm) # get the path/name of the new answer key

## data ##
datapath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Data Files"
topic = gsub('.Questions.txt','',gsub('Topic','',basename(key_txt)))
namelist = list.files(path=datapath,full.names=TRUE)
namelist = namelist[grep(paste('Topic',topic,'\\\\.',sep=''),basename(namelist))]
namelist

## learning objectives ##
LOpath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Topic Outcomes/Topic03.Outcomes.txt" ### NEW PATH!!!

## TOPIC 06 - FINAL- BE USED FOR EXAMPLE ######
##LOpath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Topic Outcomes-Final/Topic06.Outcomes.txt" ### NEW PATH!!!

## 1. Individual report ##
for (i in namelist) rewrite_data(i)
for (i in namelist) report_routine(key_txt,datafile=i,rewrite=FALSE,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-individual-short.Rnw")
for (i in namelist) report_routine(key_txt,datafile=i,rewrite=FALSE,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-individual.Rnw")

## 2. Cross-section report ##
report_routine(key_txt,namelist,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-section-short.Rnw")
report_routine(key_txt,namelist,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-section.Rnw")

## 3. Cross-topic report ##
tabfiles = set_dir(datapath)
tabfiles = tabfiles[tabfiles$section!='201' & tabfiles$topic \%in\% paste0('Topic0',1:5),]
mgdata = merge_data(tabfiles)
for (sctn in unique(tabfiles$section)) {
  merged = subset_data(mgdata,tabfiles,choice=sctn)
  knit("/Users/lindz/ePort/inst/Rnw/hw-topic.Rnw",output=paste0('Stat101hwk_Unit1_Section',sctn,'.tex'))
}

## 4. Cross-section & cross-topic report ##
merged = subset_data(mgdata,tabfiles)
knit("/Users/lindz/ePort/inst/Rnw/hw-topic-section.Rnw",output='Stat101hwk_Unit1_allSections.tex')
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

