% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCSV.R
\name{rewriteData}
\alias{rewriteData}
\title{Rewrite the data file}
\usage{
rewriteData(filenm)
}
\arguments{
\item{filenm}{file name with path}
}
\description{
Solve two problems that I used to work in gedit:
"" --> ", and "\\r\\n" --> \\r\\n.
This function will read the data, replace every two quotes ""
by one quote ", remove the quotes before and after every line,
i.e., replace "\\r\\n" by \\r\\n, then write the result to the
same file.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

