% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merge_semester.R
\name{set_dir}
\alias{set_dir}
\title{Set and parse the data directory}
\usage{
set_dir(data.dir = NULL, data.list = NULL)
}
\arguments{
\item{data.dir}{directory of the data files. Could be NULL,
then a selection GUI will pop out.}

\item{data.list}{a vector of data files with path. Suppressed
if data.dir is not NULL.}
}
\value{
a data frame of file names (with path), topics, and sections
}
\description{
All of the csv files in the directory with specific
format will be detected and recorded.
}
\examples{
library(ePort)
library(knitr)

##### Homework reports #####
## key ##
# save the .doc file as .htm format, to keep the images
key_htm = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Database Questions/Topic03.Questions.htm"
refine_key(key_htm) # generate a clean answer key with paths to the plots in the question
key_txt = gsub("htm$","txt",key_htm) # get the path/name of the new answer key

## data ##
datapath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Data Files"
topic = gsub('.Questions.txt','',gsub('Topic','',basename(key_txt)))
namelist = list.files(path=datapath,full.names=TRUE)
namelist = namelist[grep(paste('Topic',topic,'\\\\.',sep=''),basename(namelist))]
namelist

## learning objectives ##
LOpath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Topic Outcomes/Topic03.Outcomes.txt" ### NEW PATH!!!

## TOPIC 06 - FINAL- BE USED FOR EXAMPLE ######
##LOpath = "~/Dropbox/NSF Grant 2013-2015/Semesters/Spring 2015/Topic Outcomes-Final/Topic06.Outcomes.txt" ### NEW PATH!!!

## 1. Individual report ##
for (i in namelist) rewrite_data(i)
for (i in namelist) report_routine(key_txt,datafile=i,rewrite=FALSE,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-individual-short.Rnw")
for (i in namelist) report_routine(key_txt,datafile=i,rewrite=FALSE,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-individual.Rnw")

## 2. Cross-section report ##
report_routine(key_txt,namelist,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-section-short.Rnw")
report_routine(key_txt,namelist,LOfile=LOpath,knitfile="/Users/lindz/ePort/inst/Rnw/hw-section.Rnw")

## 3. Cross-topic report ##
tabfiles = set_dir(datapath)
tabfiles = tabfiles[tabfiles$section!='201' & tabfiles$topic \%in\% paste0('Topic0',1:5),]
mgdata = merge_data(tabfiles)
for (sctn in unique(tabfiles$section)) {
  merged = subset_data(mgdata,tabfiles,choice=sctn)
  knit("/Users/lindz/ePort/inst/Rnw/hw-topic.Rnw",output=paste0('Stat101hwk_Unit1_Section',sctn,'.tex'))
}

## 4. Cross-section & cross-topic report ##
merged = subset_data(mgdata,tabfiles)
knit("/Users/lindz/ePort/inst/Rnw/hw-topic-section.Rnw",output='Stat101hwk_Unit1_allSections.tex')
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

