% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSemester.R
\name{setDir}
\alias{setDir}
\title{Set and parse the data directory}
\usage{
setDir(data.dir = NULL, data.list = NULL)
}
\arguments{
\item{data.dir}{directory of the data files. Could be NULL,
then a selection GUI will pop out.}

\item{data.list}{a vector of data files with path. Suppressed 
if data.dir is not NULL.}
}
\value{
a data frame of file names (with path), topics, and sections
}
\description{
All of the csv files in the directory with specific
format will be detected and recorded.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

