% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSemester.R
\name{subsetData}
\alias{subsetData}
\title{Subset one section in use}
\usage{
subsetData(mergedat, files, choice = "all")
}
\arguments{
\item{mergedat}{output from \code{mergeData}}

\item{files}{output from \code{setDir}}

\item{choice}{section name. Default to be 'all',
which means all the sections are selected.}
}
\value{
a subset of \code{mergedat} with the same format
}
\description{
If we want to generate a report for one section with
many topics, this function could help to filter the
section from the output of \code{mergeData}.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

