% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{summaryLevel}
\alias{summaryLevel}
\title{Summary on questions}
\usage{
summaryLevel(QS, QSS, CSS = NULL)
}
\arguments{
\item{QS}{a data frame of question scores. First object of
the output from function \code{cleanScore}.}

\item{QSS}{a list. First object of the output from function
\code{cleanSet}: QuestionSet.}

\item{CSS}{a list. Second object of the output from function
\code{cleanSet}: ObjectiveSet. Could be NULL.}
}
\value{
a list of two or three objects (when \code{CSS} is not NULL).
The first two objects are the summaries by question,
and by question set. The optional third output is the
summary by learning outcome.
}
\description{
Gives the correct percentage, count, mean, standard deviation,
minimum, median, maximum, and the number of NA's
by different levels.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

