% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{summaryScore}
\alias{summaryScore}
\title{Summary on students}
\usage{
summaryScore(QS, QSS, CSS = NULL)
}
\arguments{
\item{QS}{a data frame of question scores. First object of
the output from function \code{cleanScore}.}

\item{QSS}{a list. First object of the output from function
\code{cleanSet}: QuestionSet.}

\item{CSS}{a list. Second object of the output from function
\code{cleanSet}: ObjectiveSet. Could be NULL.}
}
\value{
a data frame with total scores, scores for question sets,
learning objectives (if \code{CSS} is not NULL), and questions.
}
\description{
Merge the scores at different levels (total, learning
objectives, question sets, questions) together.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

