% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_HTML}
\alias{clean_HTML}
\title{Remove the HTML strings in the answer}
\usage{
clean_HTML(HWsheet, mark = TRUE, image = FALSE)
}
\arguments{
\item{HWsheet}{Vector or data frame of characters}

\item{mark}{Logical. If TRUE then some marks are given to the matching questions}

\item{image}{Logical. If TRUE then some marks are given to the embedded images}
}
\value{
Vector or data frame with no HTML strings
}
\description{
Remove the HTML strings in the answer
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

