% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCSV.R
\name{combineFiles}
\alias{combineFiles}
\title{Merge two CSV files into one file}
\usage{
combineFiles(file1, file2, newnm)
}
\arguments{
\item{file1}{File name with path to the first file}

\item{file2}{File name with path to the second file}

\item{newnm}{File name for the returned merged file}
}
\value{
A new CSV file
}
\description{
If one topic is organized into two chapters, then we must 
merge together the two files. Note that the Question ID and RsQ ID
are updated for the second file upon merging.
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}
\references{
Refer to package vignette Section 5.2
}

