% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportFormat.R
\name{makeReport}
\alias{makeReport}
\title{Generate a report for one homework}
\usage{
makeReport(keyFile = NULL, dataFile = NULL, loFile = NULL,
  outFile = NULL, reportType = NULL, keepFiles = FALSE, keepTex = FALSE,
  keepImage = FALSE, className = "Stat101", lowScore = 80,
  repeatLowScore = 20, topic = NULL, section = NULL, path = NULL,
  type = NULL, unit = 1, rewrite = FALSE, skip = NULL)
}
\arguments{
\item{keyFile}{file name with path for the answer key}

\item{dataFile}{the file name with path for the data}

\item{loFile}{the file name with path for the learning outcomes}

\item{outFile}{the directory to save the tex file}

\item{reportType}{the type of report}

\item{keepFiles}{keep files (.aux, .log, etc) used to generate .tex file, default value is FALSE}

\item{keepTex}{keep .tex file, default value is FALSE}

\item{keepImage}{keep individual image files used to generate .tex file, default value is FALSE}

\item{className}{the name of the class, default value is "Stat101"}

\item{lowScore}{the value that instructors designate as a minimally acceptable score (out of 100) for a given topic and section, when running the makeReport() routine with a reportType of "secTopicShort". Any student who scores below this value on their assignment will have their e-mail listed in a text file. Default value is 80}

\item{repeatLowScore}{the proportion that instructors use to determine students who have performed poorly across multiple assignments. If a student has been ranked below this proportion (out of the other students) for at least half of the assignments, then their e-mail will be listed in a text file. This parameter can be defined when running the makeReport() routine with a reportType of "crossSecUnit". Default value is 20}

\item{topic}{the topic number, could be integer or character}

\item{section}{the section name, usually one of 'AB', 'CD', 'GHQ'}

\item{path}{the directory to the data files}

\item{type}{the strings before the topic number in the csv file name}

\item{rewrite}{logical. If TRUE then the csv files are rewritten}

\item{skip}{a vector of integers. Same as the parameter in \code{cleanScore}}
}
\value{
a tex file to be compiled
}
\description{
A tex file will be generated from the specified data files,
answer key, and learning outcomes.
}
\examples{
library(ePort)
library(knitr)

##### Homework reports #####
## key ##
# save the .doc file as .htm format, to keep the images
key_htm = "C:/Users/amyf/Dropbox/Stat 101 Coordinating Materials/Online Homework Assessment Data/Questions/Topic03.Questions.htm"
refineKey(key_htm) # generate a clean answer key with paths to the plots in the question
keyPath = gsub("htm$","txt",key_htm) # get the path/name of the new answer key

## data ##
dataPath = "C:/Users/amyf/Dropbox/Stat 101 Coordinating Materials/Online Homework Assessment Data/Topic03Data"
topic = gsub('.Questions.txt','',gsub('Topic','',basename(keyFile)))
namelist = list.files(path=dataPath,full.names=TRUE)
namelist = namelist[grep(paste('Topic',topic,'\\\\.',sep=''),basename(namelist))]
namelist

## learning outcomes ##
loPath = "C:/Users/amyf/Dropbox/Stat 101 Coordinating Materials/Online Homework Assessment Data/Outcomes/Topic03.Outcomes.txt" 

## 1. Individual report ##
for (i in namelist) rewriteData(i)
for (i in namelist) makeReport(keyPath=keyFile,dataFile=i,rewrite=FALSE,loFile=loPath,reportType="inst/Rnw/hw-individual-short.Rnw")
for (i in namelist) makeReport(keyPath=keyFile,dataFile=i,rewrite=FALSE,loFile=loPath,reportType="inst/Rnw/hw-individual-long.Rnw")

## 2. Cross-section report ##
makeReport(keyFile=keyPath,namelist,loFile=loPath,reportType="inst/Rnw/hw-section-short.Rnw")
makeReport(keyFile=keyPath,namelist,loFile=loPath,reportType="inst/Rnw/hw-section.Rnw")

## 3. Cross-topic report ##
tabfiles = setDir(dataPath)
# Originally only included Topics 01-05
#tabfiles = tabfiles[tabfiles$section!='201' & tabfiles$topic \%in\% paste0('Topic0',1:5),]
tabfiles = tabfiles[tabfiles$section!='201',] # new syntax
mgdata = mergeData(tabfiles)
for (sctn in unique(tabfiles$section)) {
  merged = subsetData(mgdata,tabfiles,choice=sctn)
  knit("/Users/lindz/ePort/inst/Rnw/hw-topic.Rnw",output=paste0('Stat101hwk_Unit1_Section',sctn,'.tex'))
}

## 4. Cross-section & cross-topic report ##
merged = subsetData(mgdata,tabfiles)
knit("/Users/lindz/ePort/inst/Rnw/hw-topic-section.Rnw",output='Stat101hwk_Unit1_allSections.tex')
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

