% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCSV.R
\name{rewriteData}
\alias{rewriteData}
\title{Rewrite the data file}
\usage{
rewriteData(filenm)
}
\arguments{
\item{filenm}{file name with path}
}
\description{
Solve two problems that I used to work in gedit:
"" --> ", and "\\r\\n" --> \\r\\n.
This function will read the data, replace every two quotes ""
by one quote ", remove the quotes before and after every line,
i.e., replace "\\r\\n" by \\r\\n, then write the result to the
same file.
}
\examples{
library(ePort)
library(knitr)

##### Homework reports #####
## key ##
# save the .doc file as .htm format, to keep the images
key_htm = "C:/Users/amyf/Dropbox/Stat 101 Coordinating Materials/Online Homework Assessment Data/Questions/Topic03.Questions.htm"
refineKey(key_htm) # generate a clean answer key with paths to the plots in the question
keyPath = gsub("htm$","txt",key_htm) # get the path/name of the new answer key

## data ##
dataPath = "C:/Users/amyf/Dropbox/Stat 101 Coordinating Materials/Online Homework Assessment Data/Topic03Data"
topic = gsub('.Questions.txt','',gsub('Topic','',basename(keyFile)))
namelist = list.files(path=dataPath,full.names=TRUE)
namelist = namelist[grep(paste('Topic',topic,'\\\\.',sep=''),basename(namelist))]
namelist

## learning outcomes ##
loPath = "C:/Users/amyf/Dropbox/Stat 101 Coordinating Materials/Online Homework Assessment Data/Outcomes/Topic03.Outcomes.txt" 

## 1. Individual report ##
for (i in namelist) rewriteData(i)
for (i in namelist) makeReport(keyPath=keyFile,dataFile=i,rewrite=FALSE,loFile=loPath,reportType="inst/Rnw/hw-individual-short.Rnw")
for (i in namelist) makeReport(keyPath=keyFile,dataFile=i,rewrite=FALSE,loFile=loPath,reportType="inst/Rnw/hw-individual-long.Rnw")

## 2. Cross-section report ##
makeReport(keyFile=keyPath,namelist,loFile=loPath,reportType="inst/Rnw/hw-section-short.Rnw")
makeReport(keyFile=keyPath,namelist,loFile=loPath,reportType="inst/Rnw/hw-section.Rnw")

## 3. Cross-topic report ##
tabfiles = setDir(dataPath)
# Originally only included Topics 01-05
#tabfiles = tabfiles[tabfiles$section!='201' & tabfiles$topic \%in\% paste0('Topic0',1:5),]
tabfiles = tabfiles[tabfiles$section!='201',] # new syntax
mgdata = mergeData(tabfiles)
for (sctn in unique(tabfiles$section)) {
  merged = subsetData(mgdata,tabfiles,choice=sctn)
  knit("/Users/lindz/ePort/inst/Rnw/hw-topic.Rnw",output=paste0('Stat101hwk_Unit1_Section',sctn,'.tex'))
}

## 4. Cross-section & cross-topic report ##
merged = subsetData(mgdata,tabfiles)
knit("/Users/lindz/ePort/inst/Rnw/hw-topic-section.Rnw",output='Stat101hwk_Unit1_allSections.tex')
}
\author{
Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}

